api/requirements.txt
flask
cassandra-driver
pandas